feat: add support for custom target port

+++ src/bgpdump_option.h
@@ -19,6 +19,7 @@
 #ifndef __BGPDUMP_OPTION_H__
 #define __BGPDUMP_OPTION_H__
 
+#include <stdint.h>
 #define BGPDUMP_BUFSIZ_DEFAULT "16MiB"
 
 extern int quiet;
@@ -48,7 +49,7 @@ extern int nhs;
 extern struct sockaddr_in nhs_addr4;
 extern struct sockaddr_in6 nhs_addr6;
 extern int withdraw_delay;
-extern int prefix_limit;
+extern uint64_t prefix_limit;
 
 extern unsigned long long bufsiz;
 extern unsigned long long nroutes;
